vbglR3DoIOCtl: size_t/DWORD warnings.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@18452 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/common/VBoxGuestLib/VBoxGuestR3Lib.cpp
@@ -332,7 +332,7 @@ int vbglR3DoIOCtl(unsigned iFunction, void *pvData, size_t cbData)
 {
 #if defined(RT_OS_WINDOWS)
     DWORD cbReturned = 0;
-    if (!DeviceIoControl(g_hFile, iFunction, pvData, cbData, pvData, cbData, &cbReturned, NULL))
+    if (!DeviceIoControl(g_hFile, iFunction, pvData, (DWORD)cbData, pvData, (DWORD)cbData, &cbReturned, NULL))
     {
 /** @todo The passing of error codes needs to be tested and fixed (as does *all* the other hosts except for
  * OS/2).  The idea is that the VBox status codes in ring-0 should be transfered without loss down to
